// plio-api/plio/apis/SupportTicketController.ts
/* eslint-disable */

/**
 * Plio.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { createEndpoint, normalizePath } from '@1023-ventures/darri-core';

import {
    ProblemDetails,
    ProblemDetailsFromJSON,
    ProblemDetailsToJSON,
} from '../models/ProblemDetails';
import {
    SupportTicket,
    SupportTicketFromJSON,
    SupportTicketToJSON,
} from '../models/SupportTicket';

// Data Transfer Objects for individual endpoints

type DeleteUrlParams = {
    /** id  true*/
    id: string;
}

type DeleteQueryParams = {
    /**  */
    concurrency?: string;
}

type CreateBody =
    /**  */
    & SupportTicket;

type CreateUrlParams = {
    /** reload  true*/
    reload: boolean;
}

type ReadUrlParams = {
    /** id  true*/
    id: string;
}

type SparseUpdateBody =
    /**  */
    & string;

type SparseUpdateUrlParams = {
    /** id  true*/
    id: string;
    /** reload  true*/
    reload: boolean;
}

type UpdateBody =
    /**  */
    & SupportTicket;

type UpdateUrlParams = {
    /** reload  true*/
    reload: boolean;
}

/**
 * no description
 */
export const SupportTicketController = {

    /**
     */
    _delete: createEndpoint<boolean, never, DeleteUrlParams, DeleteQueryParams>({
        method: 'DELETE',
        url: normalizePath('/plio/api/SupportTicket/{id}')
    }),
    /**
     */
    create: createEndpoint<SupportTicket, CreateBody, CreateUrlParams, never>({
        method: 'POST',
        url: normalizePath('/plio/api/SupportTicket')
    }),
    /**
     */
    createNewTicket: createEndpoint<SupportTicket, never, never, never>({
        method: 'GET',
        url: normalizePath('/plio/api/SupportTicket/createNewTicket')
    }),
    /**
     */
    read: createEndpoint<SupportTicket, never, ReadUrlParams, never>({
        method: 'GET',
        url: normalizePath('/plio/api/SupportTicket/{id}')
    }),
    /**
     */
    readAll: createEndpoint<Array<SupportTicket>, never, never, never>({
        method: 'GET',
        url: normalizePath('/plio/api/SupportTicket')
    }),
    /**
     */
    sparseUpdate: createEndpoint<SupportTicket, SparseUpdateBody, SparseUpdateUrlParams, never>({
        method: 'PATCH',
        url: normalizePath('/plio/api/SupportTicket/{id}')
    }),
    /**
     */
    update: createEndpoint<SupportTicket, UpdateBody, UpdateUrlParams, never>({
        method: 'PUT',
        url: normalizePath('/plio/api/SupportTicket')
    }),
}